Created method: GetScaledMapAreaInfoV1 on class: BoundedMapArea. Also refactored the MapJobHelper and associated classes to use SizeDbl instead of SizeInt for the PosterSize, DisplaySize, etc.

+++ src/MSetExplorer/XPoc/Sampling/MapAreaInfoViewModel.cs
@@ -190,7 +190,7 @@ namespace MSetExplorer.XPoc
 			}
 		}
 
-		private double GetCoordDiff(double sampleWidth, int screenWidth, double mapWidth)
+		private double GetCoordDiff(double sampleWidth, double screenWidth, double mapWidth)
 		{
 			var rWidth = sampleWidth * screenWidth;
 			var result = mapWidth - rWidth;
